+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## unreleased
+* Add webhook sample for `GrantedPaymentMethodRevoked`
 * Add `paypal_messages` to `Dispute`
 * Add `tax_identifiers` parameter to `Customer.create` and `Customer.update`